FedMA_simulatuin
For fairness comparison, we adopt two baseline algorithms,named FedAvg and FedProx. Specifically, FedAvg is a classic algorithm in FL and has strong adaptability. FedProx can be viewed as a generalization and re-parametrization of FedAvg, and it is recognized as a very good FL algorithm. It solves the problem of non-IID data in FL to a certain extent. Due to time limited, we only performed a simple experimental simulation on the CIFAR-10 data set. More simulation results will be updated in final paper.